# Backend/openapi.yaml
openapi: 3.0.3
info:
  title: HamAmoz School Platform API
  version: 1.0.0 (v1)
  description: API نسخه MVP سامانه چندشعبه‌ای هم‌آموز
paths:
...
  /api/v1/guardians/:
    get:
      operationId: guardians_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: organization
        schema:
          type: string
          format: uuid
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - guardians
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGuardianList'
          description: ''
    post:
      operationId: guardians_create
      tags:
      - guardians
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardianRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/GuardianRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/GuardianRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardian'
          description: ''
...
  /api/v1/students/{id}/guardians/:
    post:
      operationId: students_guardians_create
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this student.
        required: true
      tags:
      - students
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StudentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StudentRequest'
        required: true
      security:
      - jwtAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: ''
...
components:
  schemas:
...
    GenderEnum:
      enum:
      - female
      - male
      type: string
      description: |-
        * `female` - دختر
        * `male` - پسر
...
    Guardian:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization:
          type: string
          format: uuid
        national_id:
          type: string
          nullable: true
          pattern: ^\d{10}$
          maxLength: 10
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        full_name:
          type: string
          readOnly: true
        phone_primary:
          type: string
          maxLength: 30
        phone_secondary:
          type: string
          maxLength: 30
        email:
          type: string
          format: email
          maxLength: 254
        address:
          type: string
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentGuardian'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - first_name
      - full_name
      - id
      - last_name
      - organization
      - phone_primary
      - students
      - updated_at
    GuardianRequest:
      type: object
      properties:
        organization:
          type: string
          format: uuid
        national_id:
          type: string
          nullable: true
          pattern: ^\d{10}$
          maxLength: 10
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        phone_primary:
          type: string
          minLength: 1
          maxLength: 30
        phone_secondary:
          type: string
          maxLength: 30
        email:
          type: string
          format: email
          maxLength: 254
        address:
          type: string
      required:
      - first_name
      - last_name
      - organization
      - phone_primary
...
    PaginatedGuardianList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Guardian'
...
    RelationshipEnum:
      enum:
      - father
      - mother
      - guardian
      - other
      type: string
      description: |-
        * `father` - پدر
        * `mother` - مادر
        * `guardian` - سرپرست
        * `other` - سایر
...
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        organization:
          type: string
          format: uuid
        national_id:
          type: string
          pattern: ^\d{10}$
          maxLength: 10
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        full_name:
          type: string
          readOnly: true
        birth_date:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/GenderEnum'
        status:
          $ref: '#/components/schemas/StudentStatusEnum'
        photo:
          type: string
          format: uri
        notes:
          type: string
        guardians:
          type: array
          items:
            $ref: '#/components/schemas/StudentGuardian'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - birth_date
      - created_at
      - first_name
      - full_name
      - gender
      - guardians
      - id
      - last_name
      - national_id
      - organization
      - updated_at
    StudentGuardian:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        guardian:
          type: string
          format: uuid
        guardian_name:
          type: string
          readOnly: true
        relationship:
          $ref: '#/components/schemas/RelationshipEnum'
        is_primary:
          type: boolean
        can_pick_up:
          type: boolean
      required:
      - guardian
      - guardian_name
      - id
      - relationship
...
    StudentRequest:
      type: object
      properties:
        organization:
          type: string
          format: uuid
        national_id:
          type: string
          minLength: 1
          pattern: ^\d{10}$
          maxLength: 10
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        birth_date:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/GenderEnum'
        status:
          $ref: '#/components/schemas/StudentStatusEnum'
        photo:
          type: string
          format: binary
        notes:
          type: string
      required:
      - birth_date
      - first_name
      - gender
      - last_name
      - national_id
      - organization
    StudentStatusEnum:
      enum:
      - active
      - graduated
      - transferred
      - withdrawn
      type: string
      description: |-
        * `active` - فعال
        * `graduated` - فارغ‌التحصیل
        * `transferred` - انتقالی
        * `withdrawn` - ترک‌تحصیل
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT